requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.5.7
boto3==1.17.73
botocore==1.20.73; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
click==8.0.0; python_version >= '3.6'
flask-cors==3.0.10
flask==2.0.0
gunicorn==20.1.0
itsdangerous==2.0.0; python_version >= '3.6'
jinja2==3.0.0; python_version >= '3.6'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
mako==1.1.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
markdown==3.3.4
markupsafe==2.0.0; python_version >= '3.6'
marshmallow==3.12.1
pillow==8.2.0
psycopg2==2.8.6
pyjwt==2.1.0
python-dateutil==2.8.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-editor==1.0.4
s3transfer==0.4.2
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==1.3.23
urllib3==1.26.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
werkzeug==2.0.0; python_version >= '3.6'
